로컬 미사일 위치가 갑자기 튀는 현상 수정
- 패킷에 포함된 미사일 위치 정보가 int16_t 자료형으로 구성되어 있어,
미사일 속도와, 수명을 조절하여 예상 범위를 초과하여 음수가 되어 발생하는 문제
- int32_t 자료형으로 수정하여 해결.

- 앞서 완성한 디버깅 박스로 문제 쉽게 식별함.

+++ NGP_TermProject/NGPTermProject/CSPacket.h
@@ -52,9 +52,9 @@ struct PlayerKeyPacket {
 };
 
 struct ItemInfoPacket {
-	int16_t positionX;
-	int16_t positionY;
-	int16_t positionZ;
+	int32_t positionX;
+	int32_t positionY;
+	int32_t positionZ;
 	bool active;
 };
 
@@ -67,9 +67,9 @@ struct ItemInfoBundlePacket
 struct MissileInfoPacket {
 
 	bool active;
-	int16_t positionX;  
-	int16_t positionY; 
-	int16_t positionZ; 
+	int32_t positionX;  
+	int32_t positionY; 
+	int32_t positionZ; 
 };
 
 struct MissileInfoBundlePacket
